remove most pragmas that was disabling warnings fix some warnings

+++ UnitTests/UnitTest_RotatingSkin/DemoKeeper.cpp
@@ -51,7 +51,7 @@ namespace demo
 
 	void DemoKeeper::injectMouseMove(int _absx, int _absy, int _absz)
 	{
-		rotato->setAngle(atan2((double)_absx - rotato->getCenter(false).left, -(double)_absy + rotato->getCenter(false).top));
+		rotato->setAngle(atan2((float)_absx - rotato->getCenter(false).left, -(float)_absy + rotato->getCenter(false).top));
 
 		base::BaseManager::injectMouseMove(_absx, _absy, _absz);
 	}
